added unverified Email tests including one that runs the cleanUp script
